age_of_gold_worker/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "age-of-gold-worker"
version = "1.0.0"
description = "The worker for the Age of Gold project to create default avatars and to send emails."
authors = [{ name = "Grabot" }]
license = "GPL-3.0-or-later"
requires-python = ">=3.14.1"


dependencies = [
    "celery>=5.5.3,<6",
    "pydantic-settings>=2.11.0,<3",
    "pillow>=12.0.0,<13",
    "asyncio-redis>=0.16.0,<0.17",
    "redis>=7.0.0,<8",
    "boto3>=1.42.14",
    "botocore>=1.42.14",
    "cryptography>=46.0.3,<47",
]

[tool.hatch.build.targets.sdist]
include = ["."]

[tool.hatch.build.targets.wheel]
include = ["."]
